[SAFE-72]

+++ schema/mutation/deleteStep.ts
@@ -1,11 +1,9 @@
 import { GraphQLNonNull, GraphQLID, GraphQLError } from "graphql";
 import errors from "../../const/errors";
-import permissions from "../../const/permissions";
 import deleteContent from "../../services/deleteContent";
-import checkPermission from "../../utils/checkPermission";
 import { StepType } from "../types";
-import mongoose from 'mongoose';
 import { Workflow, Step } from "../../models";
+import { AppAbility } from "../../security/defineAbilityFor";
 
 export default {
     /*  Delete a step from its id and erase its reference in the corresponding workflow.
@@ -17,23 +15,15 @@ export default {
         id: { type: new GraphQLNonNull(GraphQLID) }
     },
     async resolve(parent, args, context) {
+        // Authentication check
         const user = context.user;
+        if (!user) { throw new GraphQLError(errors.userNotLogged); }
+
+        const ability: AppAbility = context.user.ability;
         const workflow = await Workflow.findOne({ steps: args.id });
-        if (!workflow) throw new GraphQLError(errors.dataNotFound);
-        let step;
-        if (checkPermission(user, permissions.canManageApplications)) {
-            step = await Step.findByIdAndDelete(args.id);
-        } else {
-            // Check if we can update workflow and delete step
-            // if (workflow.permissions.canUpdate.some(x => context.user.roles.map(x => mongoose.Types.ObjectId(x._id)).includes(x))) {
-            // }
-            const filters = {
-                'permissions.canDelete': { $in: context.user.roles.map(x => mongoose.Types.ObjectId(x._id)) },
-                _id: args.id
-            };
-            step = await Step.findOneAndDelete(filters);
-        }
-        if (!step) throw new GraphQLError(errors.permissionNotGranted);
+        const filters = Step.accessibleBy(ability, 'delete').where({_id: args.id}).getFilter();
+        const step = await Step.findOneAndDelete(filters);
+        if (!step || !workflow) throw new GraphQLError(errors.permissionNotGranted);
         await deleteContent(step);
         const update = {
             modifiedAt: new Date(),
